+++ BertSpeak/root/src/main/kotlin/chuckcoughlin/bertspeak/data/Point2D.kt
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024. Charles Coughlin. All Rights Reserved.
+ * Copyright 2024-2025. Charles Coughlin. All Rights Reserved.
  * MIT License.
  */
 package chuckcoughlin.bertspeak.data
@@ -7,7 +7,7 @@ package chuckcoughlin.bertspeak.data
 /**
  * Use our own class of Point to make it easier to serialize and deserialize.
  */
-data class Point2D (val x1:Float, val x2:Float) {
-    val x: Float = x1
-    val y: Float = x2
+data class Point2D (val x1:Double, val x2:Double) {
+    val x: Double = x1
+    val y: Double = x2
 }
