+++ BertSpeak/root/src/main/kotlin/chuckcoughlin/bertspeak/data/Point3D.kt
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024. Charles Coughlin. All Rights Reserved.
+ * Copyright 2024-2025. Charles Coughlin. All Rights Reserved.
  * MIT License.
  */
 package chuckcoughlin.bertspeak.data
@@ -7,8 +7,8 @@ package chuckcoughlin.bertspeak.data
 /**
  * Use our own class of Point to make it easier to serialize and deserialize.
  */
-data class Point3D (val x1:Float, val x2:Float,val x3:Float) {
-    val x: Float = x1
-    val y: Float = x2
-    val z: Float = x3
+data class Point3D (val x1:Double, val x2:Double,val x3:Double) {
+    val x: Double = x1
+    val y: Double = x2
+    val z: Double = x3
 }
